Update bounded_buffer_test.c
There seems to be a difference between how Ubuntu and Mac handles the creation of new struct variables. When running the old test on a Ubuntu machine (private and ThinClient) init_test() cannot fail due to new struct variables not being set to NULL upon declaration. This causes errors in the init_test() to look like they occur in destroy_test() which is confusing for the students. Setting all of the variables to NULL provides desired behavior on Ubuntu machines. Two additional assertions added to the init_test() to check that the "in" and "out" is set on init.